orders/requirements.txt
amqp==2.5.1
billiard==3.6.1.0
celery==4.3.0
Django==2.2.5
djangorestframework==3.10.3
importlib-metadata==0.23
kombu==4.6.4
more-itertools==7.2.0
pkg-resources==0.0.0
pytz==2019.2
PyYAML==5.1.2
redis==3.3.8
sqlparse==0.3.0
vine==1.3.0
zipp==0.6.0
